pyproject.toml
[tool.poetry]
name = "python-telegram-bot_API-BF"
version = "0.1.0"
description = ""
authors = ["Stanis96 <vasilievstas96@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
flake8 = "^5.0.4"
black = "^22.10.0"
python-telegram-bot = "^13.13"
virtualenv = "^20.15.1"
django-environ = "^0.9.0"
decouple = "^0.0.7"
config = "^0.5.1"
requests = "^2.28.1"


[tool.poetry.dev-dependencies]
black = "22.10.0"
flake8 = "5.0.4"
pyproject-flake8 = "5.0.4.post1"
flake8-black = "0.3.3"
flake8-builtins = "2.0.0"
flake8-import-order = "0.18.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
exclude = """
    ('
    |.idea/
    |.venv/
    |.vscode/
    ')
"""

# call using pflake8
[tool.flake8]
max-line-length = 100
max-complexity = 4
ignore = "A003, W503"
exclude = """
    .idea/
    .venv/
    .vscode/
    */migrations/
"""

[virtualenvs]
create = true
in-project = true
path = ".venv"
